moved GetEthAssetAmount from pcommon to xmrmaker
