+++ protocol/xmrmaker/ethereum_asset_amount.go
@@ -1,7 +1,7 @@
 // Copyright 2023 The AthanorLabs/atomic-swap Authors
 // SPDX-License-Identifier: LGPL-3.0-only
 
-package protocol
+package xmrmaker
 
 import (
 	"context"
@@ -14,9 +14,9 @@ import (
 	"github.com/athanorlabs/atomic-swap/ethereum/extethclient"
 )
 
-// GetEthAssetAmount converts the passed asset amt (in standard units) to
+// getEthAssetAmount converts the passed asset amt (in standard units) to
 // EthAssetAmount (ie WeiAmount or ERC20TokenAmount)
-func GetEthAssetAmount(
+func getEthAssetAmount(
 	ctx context.Context,
 	ec extethclient.EthClient,
 	amt *apd.Decimal, // in standard units
@@ -35,8 +35,6 @@ func GetEthAssetAmount(
 		return coins.NewTokenAmountFromDecimals(amt, token), nil
 	}
 
-	// The caller had the info to check this and probably did, but we do it
-	// again here for symmetry.
 	if coins.ExceedsDecimals(amt, coins.NumEtherDecimals) {
 		return nil, fmt.Errorf("value can not be represented in ETH's %d decimals", coins.NumEtherDecimals)
 	}
